Move ordered_set and _OrderedSet to ordered_set.py
