+++ src/torchjd/autojac/_transform/ordered_set.py
@@ -1,0 +1,10 @@
+from collections import OrderedDict
+from typing import Iterable, TypeAlias
+
+from torchjd.autojac._transform._utils import _KeyType
+
+_OrderedSet: TypeAlias = OrderedDict[_KeyType, None]
+
+
+def ordered_set(elements: Iterable[_KeyType]) -> _OrderedSet[_KeyType]:
+    return OrderedDict.fromkeys(elements, None)
